Refactor listing page
