+++ static/js/publisher/listing/components/App/App.tsx
@@ -8,6 +8,8 @@ import {
   Button,
 } from "@canonical/react-components";
 
+import { getChanges, getFormData, getListingData } from "../../utils";
+
 import PageHeader from "../PageHeader";
 import SaveAndPreview from "../SaveAndPreview";
 import ListingDetailsSection from "../../sections/ListingDetailsSection";
@@ -15,84 +17,14 @@ import ContactInformationSection from "../../sections/ContactInformationSection"
 import AdditionalInformationSection from "../../sections/AdditionalInformationSection";
 import PreviewForm from "../PreviewForm";
 
-interface DirtyFields {
-  [key: string]: any;
-}
-
-type License = {
-  key: string;
-  name: string;
-};
-
 function App() {
   const snapId = window?.listingData?.snap_id;
   const publisherName = window?.listingData?.publisher_name;
   const categories = window?.listingData?.categories;
 
-  const licenseSort = (a: License, b: License) => {
-    if (a.name < b.name) {
-      return -1;
-    }
-
-    if (a.name > b.name) {
-      return 1;
-    }
-
-    return 0;
-  };
-
-  const [listingData, setListingData] = useState({
-    snap_name: window?.listingData?.snap_name,
-    title: window?.listingData?.snap_title,
-    summary: window?.listingData?.summary,
-    description: window?.listingData?.description,
-    website: window?.listingData?.website,
-    contact: window?.listingData?.contact,
-    categories: window?.listingData?.categories,
-    images: [],
-    public_metrics_enabled: window?.listingData?.public_metrics_enabled,
-    public_metrics_blacklist: window?.listingData?.public_metrics_blacklist,
-    license: window?.listingData?.license,
-    license_type: window?.listingData?.license_type,
-    licenses: window?.listingData?.licenses.sort(licenseSort),
-    video_urls: window?.listingData?.video_urls[0],
-    "primary-category": window?.listingData?.snap_categories?.categories[0],
-    "secondary-category": window?.listingData?.snap_categories?.categories[1],
-    public_metrics_territories: !window?.listingData?.public_metrics_blacklist.includes(
-      "installed_base_by_country_percent"
-    ),
-    public_metrics_distros: !window?.listingData?.public_metrics_blacklist.includes(
-      "weekly_installed_base_by_operating_system_normalized"
-    ),
-    update_metadata_on_release: window?.listingData?.update_metadata_on_release,
-    contacts: window?.listingData?.links?.contact.map((link: string) => {
-      return {
-        url: link,
-      };
-    }),
-    donations: window?.listingData?.links?.donation.map((link: string) => {
-      return {
-        url: link,
-      };
-    }),
-    issues: window?.listingData?.links?.issues.map((link: string) => {
-      return {
-        url: link,
-      };
-    }),
-    "source-code": window?.listingData?.links?.["source-code"].map(
-      (link: string) => {
-        return {
-          url: link,
-        };
-      }
-    ),
-    websites: window?.listingData?.links?.website.map((link: string) => {
-      return {
-        url: link,
-      };
-    }),
-  });
+  const [listingData, setListingData] = useState(
+    getListingData(window?.listingData)
+  );
 
   const [isSaving, setIsSaving] = useState(false);
   const [hasSaved, setHasSaved] = useState(false);
@@ -118,7 +50,7 @@ function App() {
 
   const isDirty = formState.isDirty;
   const isValid = formState.isValid;
-  const dirtyFields: DirtyFields = formState.dirtyFields;
+  const dirtyFields: { [key: string]: any } = formState.dirtyFields;
 
   const onSubmit = (data: any) => {
     if (listingData?.update_metadata_on_release) {
@@ -130,88 +62,8 @@ function App() {
   };
 
   const submitForm = (data: any) => {
-    const changes: DirtyFields = {};
-    const keys = Object.keys(dirtyFields);
-    const forbiddenKeys = [
-      "primary-category",
-      "secondary-category",
-      "contacts",
-      "donations",
-      "issues",
-      "source-code",
-      "websites",
-      "licenses",
-    ];
-
-    if (
-      dirtyFields.contacts ||
-      dirtyFields.donations ||
-      dirtyFields.issues ||
-      dirtyFields.license ||
-      dirtyFields["source-code"] ||
-      dirtyFields.website
-    ) {
-      changes.links = {
-        contact: data?.contacts,
-        donation: data?.donations,
-        issues: data?.issues,
-        license: data?.license,
-        "source-code": data?.["source-code"],
-        website: data?.websites,
-      };
-    }
-
-    keys.forEach((key) => {
-      if (!forbiddenKeys.includes(key) && dirtyFields[key] === true) {
-        changes[key] = data[key];
-      }
-
-      if (
-        dirtyFields["primary-category"] ||
-        dirtyFields["secondary-category"]
-      ) {
-        changes.categories = [];
-
-        if (data["primary-category"]) {
-          changes.categories.push(data["primary-category"]);
-        }
-
-        if (data["secondary-category"]) {
-          changes.categories.push(data["secondary-category"]);
-        }
-      }
-    });
-
-    const formatLinkFields = (fields: Array<{ url: string }>) => {
-      return fields.map((item) => item.url);
-    };
-
-    const formData = new FormData();
-
-    formData.set("csrf_token", window.CSRF_TOKEN);
-    formData.set("snap_id", snapId);
-    formData.set("title", data?.title);
-    formData.set("summary", data?.summary);
-    formData.set("description", data?.description);
-    formData.set("video_urls", data?.video_urls);
-    formData.set("website", data?.website);
-    formData.set("contact", data?.contact);
-    formData.set("primary-category", data?.["primary-category"]);
-    formData.set("secondary-category", data?.["secondary-category"]);
-    formData.set("public_metrics_enabled", data?.public_metrics_enabled);
-    formData.set("public_metrics_blacklist", data?.public_metrics_blacklist);
-    formData.set("license", data?.license || "unset");
-    formData.set(
-      "links",
-      JSON.stringify({
-        contact: formatLinkFields(data?.contacts),
-        donation: formatLinkFields(data?.donations),
-        issues: formatLinkFields(data?.issues),
-        "source-code": formatLinkFields(data?.["source-code"]),
-        website: formatLinkFields(data?.websites),
-      })
-    );
-    formData.set("changes", JSON.stringify(changes));
+    const changes = getChanges(dirtyFields, data);
+    const formData = getFormData(data, snapId, changes);
 
     setIsSaving(true);
